chore(server): refactor and unify files infrastructure tests
