+++ server/internal/infrastructure/user/users_sqlite_test.go
@@ -15,7 +15,7 @@ func TestSqliteUsers_Count(t *testing.T) {
 	t.Run("returns the number of users", func(t *testing.T) {
 		t.Parallel()
 
-		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB(testutil.SqliteDBConfig{}))
+		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB())
 		defer done()
 
 		db := testutil.SqliteDBFromContext(ctx, t)
@@ -36,7 +36,7 @@ func TestSqliteUsers_Save(t *testing.T) {
 	t.Run("creates a new user", func(t *testing.T) {
 		t.Parallel()
 
-		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB(testutil.SqliteDBConfig{}))
+		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB())
 		defer done()
 		db := testutil.SqliteDBFromContext(ctx, t)
 
@@ -52,7 +52,7 @@ func TestSqliteUsers_Save(t *testing.T) {
 	t.Run("updates an existing user", func(t *testing.T) {
 		t.Parallel()
 
-		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB(testutil.SqliteDBConfig{}))
+		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB())
 		defer done()
 
 		db := testutil.SqliteDBFromContext(ctx, t)
@@ -79,7 +79,7 @@ func TestSqliteUsers_ByID(t *testing.T) {
 	t.Run("returns the correct user", func(t *testing.T) {
 		t.Parallel()
 
-		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB(testutil.SqliteDBConfig{}))
+		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB())
 		defer done()
 
 		db := testutil.SqliteDBFromContext(ctx, t)
@@ -98,7 +98,7 @@ func TestSqliteUsers_ByID(t *testing.T) {
 	t.Run("does not return a different user", func(t *testing.T) {
 		t.Parallel()
 
-		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB(testutil.SqliteDBConfig{}))
+		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB())
 		defer done()
 
 		db := testutil.SqliteDBFromContext(ctx, t)
@@ -119,7 +119,7 @@ func TestSqliteUsers_ByUsername(t *testing.T) {
 	t.Run("returns the correct user", func(t *testing.T) {
 		t.Parallel()
 
-		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB(testutil.SqliteDBConfig{}))
+		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB())
 		defer done()
 
 		db := testutil.SqliteDBFromContext(ctx, t)
@@ -138,7 +138,7 @@ func TestSqliteUsers_ByUsername(t *testing.T) {
 	t.Run("does not return a different user", func(t *testing.T) {
 		t.Parallel()
 
-		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB(testutil.SqliteDBConfig{}))
+		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithSqliteDB())
 		defer done()
 
 		db := testutil.SqliteDBFromContext(ctx, t)
