+++ server/internal/web/tests/sessions_test.go
@@ -17,7 +17,7 @@ import (
 
 func TestApi_Sessions(t *testing.T) {
 	t.Run("creates a new session for a user", func(t *testing.T) {
-		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithTempDir(), testutil.WithSqliteDB(testutil.SqliteDBConfig{}))
+		ctx, done := testutil.IntegrationTest(context.Background(), t, testutil.WithTempDir(), testutil.WithSqliteDB())
 		defer done()
 
 		app, handler := setup(ctx, t)
